+++ backend/src/main/java/de/openvalidation/openvalidationidebackend/entities/attribute/AttributeType.java
@@ -1,4 +1,4 @@
-package de.openvalidation.openvalidationidebackend.ruleset.attribute;
+package de.openvalidation.openvalidationidebackend.entities.attribute;
 
 public enum AttributeType {
   BOOLEAN,
