+++ backend/src/main/java/de/openvalidation/openvalidationidebackend/entities/ruleset/RulesetController.java
@@ -1,10 +1,10 @@
-package de.openvalidation.openvalidationidebackend.ruleset;
+package de.openvalidation.openvalidationidebackend.entities.ruleset;
 
-import de.openvalidation.openvalidationidebackend.ruleset.attribute.AttributeCreateDto;
-import de.openvalidation.openvalidationidebackend.ruleset.attribute.AttributeDto;
-import de.openvalidation.openvalidationidebackend.ruleset.attribute.AttributeUpdateDto;
-import de.openvalidation.openvalidationidebackend.ruleset.schema.SchemaDto;
-import de.openvalidation.openvalidationidebackend.ruleset.schema.SchemaUpdateDto;
+import de.openvalidation.openvalidationidebackend.entities.attribute.AttributeCreateDto;
+import de.openvalidation.openvalidationidebackend.entities.attribute.AttributeDto;
+import de.openvalidation.openvalidationidebackend.entities.attribute.AttributeUpdateDto;
+import de.openvalidation.openvalidationidebackend.entities.schema.SchemaDto;
+import de.openvalidation.openvalidationidebackend.entities.schema.SchemaUpdateDto;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
