+++ backend/src/main/java/de/openvalidation/openvalidationidebackend/entities/ruleset/RulesetNotFoundException.java
@@ -1,4 +1,4 @@
-package de.openvalidation.openvalidationidebackend.ruleset;
+package de.openvalidation.openvalidationidebackend.entities.ruleset;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
