+++ backend/src/main/java/de/openvalidation/openvalidationidebackend/entities/schema/SchemaRepository.java
@@ -1,4 +1,4 @@
-package de.openvalidation.openvalidationidebackend.ruleset.schema;
+package de.openvalidation.openvalidationidebackend.entities.schema;
 
 import org.springframework.data.jpa.repository.JpaRepository;
